Implemented JWT authentication in DiscountController
Added JWT authentication to the DiscountController. The `Microsoft.AspNetCore.Authorization` namespace was included in `DiscountController.cs` and the `Authorize` attribute was added to the `DiscountController` class. The `Microsoft.AspNetCore.Authentication.JwtBearer` package was added to `MultiShop.Discount.csproj` and used in `Program.cs` to configure the JWT authentication scheme. The `app.UseAuthentication()` middleware was added to the pipeline in `Program.cs` to enable authentication. The `IdentityServerUrl` configuration setting was added to `appsettings.json` to be used as the authority in the JWT bearer options.